# erp/docs/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERP API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://127.0.0.1:4530" }
  ],
  "paths": {
...
    "/v1/auth/admin/login": {
      "post": {
        "summary": "Admin login",
        "requestBody": {
          "required": true
        },
        "responses": {
          "200": {
            "description": "JWT token"
          }
        }
      }
    },
...
  }
}